apps/oled_piroman5/requirements.txt
luma.oled
luma.core
Pillow
gpiozero
pyserial
